Fix zoom out animation when releasing touches one by one
When zooming to a factor close to one, the image is supposed to
animate back to a fully unzoomed state soon as the pinch gesture
ends. So far, when finishing the pinch gesture by first releasing one
of the touches, the image got stuck, though. Especially when scaling
the image down to a factor < 1, this looked broken since it left the
image in a position that it normally can't end up in.

Releasing one of the touches is handled as the beginning of a drag
gesture. This is correct since it shall be possible to continue
dragging the image. Once the remaining touch is released as well,
`_handleDragEnd` [1] first calls `_end` to start the zoom out
animation, but then it calls `_realizeInertia`. When the pinch gesture
was followed by a drag gesture, velocity has been recorded and the
zoom out animation is aborted [2].

To fix the problem, we reset inertia once the need for a zoom out
animation has been detected. Notice that, if we are going to zoom back
out, there can never be a need for an inertia animation anyway.

[1] https://github.com/retyui/react-quick-pinch-zoom/blob/468eaaeb2c98bc0b26ca26b01f82b466546dbdc1/src/PinchZoom/component.tsx#L260-L261

[2] https://github.com/retyui/react-quick-pinch-zoom/blob/468eaaeb2c98bc0b26ca26b01f82b466546dbdc1/src/PinchZoom/component.tsx#L216