Add prop to prevent moving child that is smaller than container
So far, when external styles force the container to be larger than the
child, `_sanitizeOffset` allowed moving the child inside the
container.

Especially on portrait screens it can make sense to use a viewport
sized container. A landscape image will initially be centered and
scaled to fit into the container. Once the image is zoomed the whole
viewport can be used to view the image.

As long as the image still fits into the viewport vertically, it does
not make sense to allow moving it vertically. There already is a prop
`draggableUnZoomed` which prevents dragging the image while it has not
been zoomed. Soon as the image has only slightly been zoomed, it can
again be moved out of the center vertically.

This commit adds a `centerContained` prop to allow ensuring that
images remain centered as long as they still fit into the container.

The tests for `getOffsetBounds` make sure that the new implementation
behaves the same when the prop is not set.