Improve double tap zoom when container is larger than child
When using a container that is larger than the child (e.g., to build
an image viewer that uses the whole viewport on portrait screens, but
displays in a centered "contain" position initially), double tapping
outside of the image, zoomed to an "insane" offset.

To prevent this, we sanitize the offset in each step of the
animation to ensure. In particular, in combination with the previous
commit, the image remains in a centered positioned until it exceeds
the viewport dimension and only then starts to change offset.

For cases, where the container does not exceed the child size, this
commit does not change anything since each intermediate offset will be
"sane" anyway.